Add time series DSL module (#1904)
* Add time series DSL module
* Remove code duplication
