--- time-series/time-series-api/src/main/resources/META-INF/groovy/org.codehaus.groovy.runtime.ExtensionModule
@@ -1,3 +1,0 @@
-moduleName = powsybl-node-calc-module
-moduleVersion = 1.0
-extensionClasses = com.powsybl.timeseries.NodeCalcGroovyExtensionModule
