+++ time-series/time-series-api/src/test/java/com/powsybl/timeseries/DoubleTimeSeriesTest.java
@@ -9,7 +9,6 @@ package com.powsybl.timeseries;
 import com.google.common.collect.ImmutableList;
 import org.junit.Test;
 
-import java.time.Instant;
 import java.util.Iterator;
 
 import static org.junit.Assert.*;
@@ -67,14 +66,5 @@ public class DoubleTimeSeriesTest {
         assertEquals(4d, point6.getValue(2), 0d);
 
         assertFalse(it.hasNext());
-    }
-
-    @Test
-    public void builderTest() {
-        DoubleTimeSeries a = TimeSeries.createDouble("a", IrregularTimeSeriesIndex.create(Instant.now()), 5d);
-        DoubleTimeSeries b = DoubleTimeSeries.fromTimeSeries(a)
-                                             .build("ts['b'] = ts['a'] + 1")
-                                             .get(0);
-        assertArrayEquals(new double[] {6d}, b.toArray(), 0d);
     }
 }
